share index-range bound key construction
