+++ crates/jazz-tools/src/storage/surrealkv.rs
@@ -31,10 +31,11 @@ use crate::query_manager::types::Value;
 use crate::sync_manager::PersistenceTier;
 
 use super::{
-    LoadedBranch, Storage, StorageError, encode_value,
+    LoadedBranch, Storage, StorageError,
     key_codec::{
-        ack_key, branch_tips_key, commit_key, commit_prefix, increment_bytes, increment_string,
-        index_entry_key, index_prefix, index_value_prefix, obj_meta_key, parse_uuid_from_index_key,
+        ack_key, branch_tips_key, commit_key, commit_prefix, increment_bytes, index_entry_key,
+        index_prefix, index_range_scan_bounds, index_value_prefix, obj_meta_key,
+        parse_uuid_from_index_key,
     },
 };
 
@@ -461,45 +462,10 @@ impl Storage for SurrealKvStorage {
         start: Bound<&Value>,
         end: Bound<&Value>,
     ) -> Vec<ObjectId> {
-        let base_prefix = index_prefix(table, column, branch);
-
-        // Compute start key
-        let start_key = match start {
-            Bound::Included(v) => {
-                format!("{}{}", base_prefix, hex::encode(encode_value(v)))
-            }
-            Bound::Excluded(v) => {
-                let encoded = hex::encode(encode_value(v));
-                let mut key = format!("{}{}:", base_prefix, encoded);
-                // After the last possible entry for this value.
-                increment_string(&mut key);
-                key
-            }
-            Bound::Unbounded => base_prefix.clone(),
-        };
-
-        // Compute end key
-        let end_key = match end {
-            Bound::Included(v) => {
-                let encoded = hex::encode(encode_value(v));
-                let mut key = format!("{}{}:", base_prefix, encoded);
-                // Include all entries with this value by going past last UUID.
-                increment_string(&mut key);
-                key
-            }
-            Bound::Excluded(v) => {
-                format!("{}{}", base_prefix, hex::encode(encode_value(v)))
-            }
-            Bound::Unbounded => {
-                let mut end = base_prefix.clone();
-                increment_string(&mut end);
-                end
-            }
-        };
-
-        if start_key >= end_key {
+        let Some((start_key, end_key)) = index_range_scan_bounds(table, column, branch, start, end)
+        else {
             return Vec::new();
-        }
+        };
 
         let Ok(txn) = self.begin_read_txn() else {
             return Vec::new();
